pyproject.toml
[tool.poetry]
name = "spotify-playlists-creator"
version = "0.1.0"
description = "Spotify playlists creator"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
streamlit = "^1.42.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
